Rebase reconstruction to native cluster and track classes + related changes

+++ test/runVTTrackFinderCA.C
@@ -1,12 +1,12 @@
 /*
   CA Track Finder Macro
-  
+
   Usage:
     root -l
     root [0] .L setup_macros.C     // Load NA6PRoot libraries and setup paths
     root [1] .L runTrackFinderCA.C+  // Compile the macro with ACLiC
     root [2] runTrackFinderCA()      // Run with default parameters
-    
+
   Or with custom parameters:
     root [2] runTrackFinderCA(0, 100, "../fullgeo", 5)
 */
@@ -29,9 +29,10 @@
 const int maxIterationsCA = NA6PTrackerCA::kMaxIterationsCA;
 
 void runVTTrackFinderCA(int firstEv = 0,
-		      int lastEv = 9999,
-		      const char *dirSimu = "../testN6Proot/pions/latesttag",
-		      int nLayers = 5){
+                        int lastEv = 9999,
+                        const char* dirSimu = "../testN6Proot/pions/latesttag",
+                        int nLayers = 5)
+{
 
   auto magField = new MagneticField();
   magField->loadField();
@@ -89,7 +90,7 @@ void runVTTrackFinderCA(int firstEv = 0,
     mcTree->GetEvent(jEv);
     th->GetEvent(jEv);
     int nPart = mcArr->size();
-    double zvert = 0;
+    float zvert = 0;
     // get primary vertex position from the Kine Tree
     for (int jp = 0; jp < nPart; jp++) {
       auto curPart = mcArr->at(jp);
@@ -112,13 +113,13 @@ void runVTTrackFinderCA(int firstEv = 0,
         }
       }
       if (maskHits == (1 << nLayers) - 1) {
-        double pxPart = curPart.Px();
-        double pyPart = curPart.Py();
-        double pzPart = curPart.Pz();
-        double momPart = curPart.P();
-        double phiPart = curPart.Phi();
-        double thetaPart = std::acos(pzPart / momPart);
-        double etaPart = -std::log(std::tan(thetaPart / 2.));
+        float pxPart = curPart.Px();
+        float pyPart = curPart.Py();
+        float pzPart = curPart.Pz();
+        float momPart = curPart.P();
+        float phiPart = curPart.Phi();
+        float thetaPart = std::acos(pzPart / momPart);
+        float etaPart = -std::log(std::tan(thetaPart / 2.));
         hMomGen->Fill(momPart);
         hEtaGen->Fill(etaPart);
       }
@@ -133,13 +134,13 @@ void runVTTrackFinderCA(int firstEv = 0,
       int jIteration = tr.getCAIteration();
       if (tr.getNHits() == 5) {
         auto curPart = mcArr->at(std::abs(idPartTrack));
-        double pxPart = curPart.Px();
-        double pyPart = curPart.Py();
-        double pzPart = curPart.Pz();
-        double momPart = curPart.P();
-        double phiPart = curPart.Phi();
-        double thetaPart = std::acos(pzPart / momPart);
-        double etaPart = -std::log(std::tan(thetaPart / 2.));
+        float pxPart = curPart.Px();
+        float pyPart = curPart.Py();
+        float pzPart = curPart.Pz();
+        float momPart = curPart.P();
+        float phiPart = curPart.Phi();
+        float thetaPart = std::acos(pzPart / momPart);
+        float etaPart = -std::log(std::tan(thetaPart / 2.));
         hMomReco->Fill(momPart);
         hEtaReco->Fill(etaPart);
         if (jIteration >= 0 && jIteration < maxIterationsCA) {
@@ -157,14 +158,14 @@ void runVTTrackFinderCA(int firstEv = 0,
   TH1F* hPurityMom = (TH1F*)hMomGoodReco->Clone("hPurityMom");
   hPurityMom->GetYaxis()->SetTitle("purity");
   for (int iBin = 1; iBin <= hMomGoodReco->GetNbinsX(); iBin++) {
-    double cg = hMomGoodReco->GetBinContent(iBin);
-    double ct = hMomReco->GetBinContent(iBin);
+    float cg = hMomGoodReco->GetBinContent(iBin);
+    float ct = hMomReco->GetBinContent(iBin);
     if (ct == 0) {
       hPurityMom->SetBinContent(iBin, 0.);
       hPurityMom->SetBinError(iBin, 0.);
     } else {
-      double p = cg / ct;
-      double ep = std::sqrt(p * (1 - p) / ct);
+      float p = cg / ct;
+      float ep = std::sqrt(p * (1 - p) / ct);
       hPurityMom->SetBinContent(iBin, p);
       hPurityMom->SetBinError(iBin, ep);
     }
@@ -172,14 +173,14 @@ void runVTTrackFinderCA(int firstEv = 0,
   TH1F* hPurityEta = (TH1F*)hEtaGoodReco->Clone("hPurityEta");
   hPurityEta->GetYaxis()->SetTitle("purity");
   for (int iBin = 1; iBin <= hEtaGoodReco->GetNbinsX(); iBin++) {
-    double cg = hEtaGoodReco->GetBinContent(iBin);
-    double ct = hEtaReco->GetBinContent(iBin);
+    float cg = hEtaGoodReco->GetBinContent(iBin);
+    float ct = hEtaReco->GetBinContent(iBin);
     if (ct == 0) {
       hPurityEta->SetBinContent(iBin, 0.);
       hPurityEta->SetBinError(iBin, 0.);
     } else {
-      double p = cg / ct;
-      double ep = std::sqrt(p * (1 - p) / ct);
+      float p = cg / ct;
+      float ep = std::sqrt(p * (1 - p) / ct);
       hPurityEta->SetBinContent(iBin, p);
       hPurityEta->SetBinError(iBin, ep);
     }
